+++ pkg/pjrt/cpuversion.go
@@ -1,8 +1,10 @@
 package pjrt
 
+import "github.com/gomlx/go-xla/internal/utils"
+
 // DefaultCPUVersion is the default version of the CPU PJRT plugin to use with
 // this go-xla release, against which it was tested.
 //
 // See:
 // https://github.com/gomlx/pjrt-cpu-binaries/releases/tag/v0.83.1
-var DefaultCPUVersion = "v0.83.1"
+var DefaultCPUVersion = utils.DefaultCPUVersion