+++ docs/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.1.4
+
+* Updated installer library/cli to support linux/arm64 and windows/amd64. 
+  Generalized CPU installation.
+* linux and amazonlinux now use the same PJRT binary, built on a glibc-2.35 system (Ubuntu 22.04).
+* Updated dependency to `pjrt-cpu-binaries` v0.83.3.
+
 # v0.1.3
 
 * Replaced GenPool by the `internal/pool.Pool`: it simplifies a bit, a bit faster and one less